refactor(context): extract bundle render helpers
